Move scripts to new `lib` folder
Thanks @thisandagain for the tip!
