+++ lib/locale-compare.js
@@ -46,7 +46,7 @@ Helpers.mergeNewTranslations = function (existingTranslations, newTranslations, 
  *
  * @param   {object}    idICUMap    map where key=icuString, value=react-intl id
  *
- * @return  {objec}
+ * @return  {object}
  */
 Helpers.getMD5Map = function (ICUIdMap) {
     var md5Map = {};
