+++ test/fixtures/test_es_md5map.json
@@ -1,0 +1,4 @@
+{
+    "2ec20d41b181e1a41c071e13f414a74d": "test.id1",
+    "37ba6d5ef524504215f478912155f9ba": "test.id2"
+}
