+++ test/functional/build_locales_complex_strings.js
@@ -3,12 +3,11 @@ var path = require('path');
 var po2icu = require('po2icu');
 var tap = require('tap');
 
-var buildLocales = require('../../src/scripts/buildLocales/helpers');
+var buildLocales = require('../../lib/locale-compare');
 
 tap.test('buildLocalesFile', function (t) {
-    var templates = JSON.parse(fs.readFileSync(path.resolve(__dirname, '../fixtures/build_locales.json'), 'utf8'));
-    var md5map = buildLocales.getMD5Map(templates);
-    var newTranslations = po2icu.poFileToICUSync('es', path.resolve(__dirname, '../fixtures/build_locales.po'));
+    var md5map = JSON.parse(fs.readFileSync(path.resolve(__dirname, '../fixtures/test_es_md5map.json'), 'utf8'));
+    var newTranslations = po2icu.poFileToICUSync('es', path.resolve(__dirname, '../fixtures/test_es.po'));
     var translations = buildLocales.mergeNewTranslations({}, newTranslations, md5map);
 
     t.ok(translations['test.id1'] !== undefined);
