+++ test/functional/build_locales_mergeTranslations.js
@@ -1,24 +1,20 @@
-var crypto = require('crypto');
 var tap = require('tap');
 
-var buildLocales = require('../../src/scripts/buildLocales/helpers');
+var buildLocales = require('../../lib/locale-compare');
 
 tap.test('buildLocalesMergeTranslations', function (t) {
     var existingTranslations = {
         'test.test1': 'It\'s like raaayaaain, on your wedding day',
         'test.test2': 'Free to flyyy, when you already paid'
     };
-    var md5Test1 = crypto.createHash('md5').update(existingTranslations['test.test1'], 'utf8').digest('hex');
-    var md5Test2 = crypto.createHash('md5').update(existingTranslations['test.test2'], 'utf8').digest('hex');
-    var md5map = {};
-    md5map[md5Test1] = 'test.test1';
-    md5map[md5Test2] = 'test.test2';
-
     var newTranslations = {
         'Isn\'t it ironic? No.': 'Es irónico? No.'
     };
-    var md5Test3 = crypto.createHash('md5').update('Isn\'titironic?No.', 'utf8').digest('hex');
-    md5map[md5Test3] = 'test.test3';
+    var md5map = {
+        'c21ce5ceefe167028182032d4255a384': 'test.test1',
+        '9c40648034e467e16f8d6ae24bd610ab': 'test.test2',
+        '6885a345adafb3a9dd43d9f549430c88': 'test.test3'
+    };
 
     var mergedTranslations = buildLocales.mergeNewTranslations(existingTranslations, newTranslations, md5map);
     t.ok(mergedTranslations['test.test3'] !== undefined);
